// src/types/partials/GraphDefinitionLink.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TInt } from '../simpleTypes/Int';
import { TGraphDefinitionTarget } from '../partials/GraphDefinitionTarget';

export type TGraphDefinitionLink = {
    id?: String;
    extension?: TExtension[];
    modifierExtension?: TExtension[];
    path?: String;
    sliceName?: String;
    min?: TInt;
    max?: String;
    description?: String;
    target?: TGraphDefinitionTarget[];
};